test(bdd_config): refactor configuration BDD tests into helper module
- Moved extensive BDD configuration step implementations from `tests/bdd_config.rs` into a new
  helper module `tests/bdd_config_helpers.rs`.
- Updated `bdd_config.rs` to re-export the helper items and reduce clutter.
- This improves code organization and maintainability of configuration tests.
- Minor docs fixes in `docs/execplans/1-3-3-agent-and-workspace-config.md` and doc comments.
